#Testing Fragments

+++ app/src/main/java/com/test/tddandroid/database/ShoppingItem.kt
@@ -4,4 +4,4 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "shopping_item")
-data class ShoppingItem(var name:String, var amount:Int, var price:Int, var imageUrl:String, @PrimaryKey(autoGenerate = true) val id:Int?=null)
+data class ShoppingItem(var name:String, var amount:Int, var price:Float, var imageUrl:String, @PrimaryKey(autoGenerate = true) val id:Int?=null)
